+++ moPepGen/cli/__main__.py
@@ -27,7 +27,7 @@ CLI_MAIN_DESCRIPTION = """
    splitDatabase       Split noncanonical peptides into separate databases.
    mergeFasta          Merge multiple variant peptide FASTA databases.
    encodeFasta         Encode variant peptide FASTA file header.
-   decoyDatabase       Generate decoy database FASTA file.
+   decoyFasta          Generate decoy database FASTA file.
 """
 
 def main():
@@ -64,7 +64,7 @@ def main():
     cli.add_subparser_index_gvf(subparsers)
     cli.add_subparser_merge_fasta(subparsers)
     cli.add_subparser_encode_fasta(subparsers)
-    cli.add_subparser_decoy_database(subparsers)
+    cli.add_subparser_decoy_fasta(subparsers)
 
     # allowing values to start with -, such as -100,3
     # https://stackoverflow.com/a/21446783/11081630